+++ src/App.js
@@ -1,17 +1,17 @@
 import * as React from 'react';
 // import { connect } from 'react-redux';
 import Category from './Category/Category'; 
-import initMock from './mock/categories';
+import AxosMock from './mock/AxiosMock';
 
 if (window.location.host === 'localhost:3000') {
-  initMock()
+  AxosMock.initMock()
 }
 
 export default class App extends React.Component {
   render() {
     return (
       <div className="App">
-        <Category categories={[]} />
+        <Category />
       </div>
     );
   }
